+++ Data/Metrology/NoVector/Poly.hs
@@ -1,33 +1,24 @@
-{- Data/Metrology/Poly.hs
-
-   The units Package
-   Copyright (c) 2013 Richard Eisenberg
-   eir@cis.upenn.edu
-
-   This file gathers and exports all user-visible pieces of the units package.
-   It also defines the main creators and consumers of dimensioned objects.
--}
-
-{-# LANGUAGE ExplicitNamespaces, DataKinds, FlexibleInstances, TypeFamilies,
-             TypeOperators, ConstraintKinds, ScopedTypeVariables,
-             FlexibleContexts, UndecidableInstances #-}
-
 -----------------------------------------------------------------------------
 -- |
--- Module      :  Data.Metrology.Poly
--- Copyright   :  (C) 2013 Richard Eisenberg
+-- Module      :  Data.Metrology.NoVector.Poly
+-- Copyright   :  (C) 2014 Richard Eisenberg
 -- License     :  BSD-style (see LICENSE)
 -- Maintainer  :  Richard Eisenberg (eir@cis.upenn.edu)
 -- Stability   :  experimental
 -- Portability :  non-portable
 --
--- This module exports all the gubbins needed for type-checking your
--- dimensioned quantities. See 'Data.Metrology' for some functions
--- restricted to using a default LCSU, which is suitable for many
--- applications.
+-- This file exports the same definitions as Data.Metrology.Poly, but it
+-- restricts arithmetic operators to just use the Prelude numerical hierarchy,
+-- instead of vector-space. It uses orphan instances for vector-space classes
+-- to do this, so this module is not generally suitable for use in a released
+-- library. On the other hand, using these definitions may be convenient for
+-- quick experimentation, because the vector-space classes interfere with
+-- GHC's ambiguity-resolution for overloaded numbers.
 -----------------------------------------------------------------------------
 
-module Data.Metrology.Poly (
+{-# LANGUAGE TypeOperators, ConstraintKinds, ScopedTypeVariables, DataKinds #-}
+
+module Data.Metrology.NoVector.Poly (
   -- * Term-level combinators
 
   -- | The term-level arithmetic operators are defined by
@@ -37,10 +28,10 @@ module Data.Metrology.Poly (
   -- ** Additive operations
   zero, (|+|), (|-|), qSum, qNegate,
 
-  -- ** Multiplicative operations between non-vector quantities
+  -- ** Multiplicative operations between quantities
   (|*|), (|/|), (/|),
 
-  -- ** Multiplicative operations between a vector and a scalar
+  -- ** Multiplicative operations between a quantity and a non-quantity
   (*|), (|*), (|/),
 
   -- ** Exponentiation
@@ -51,7 +42,7 @@ module Data.Metrology.Poly (
   qCompare, (|<|), (|>|), (|<=|), (|>=|), (|==|), (|/=|),
   qApprox, qNapprox,        
 
-  -- * Nondimensional units, conversion between quantities and numeric values
+  -- * Conversion between quantities and numeric values
   numIn, (#), quOf, (%), showIn,
   unity, redim, convert,
   defaultLCSU, constant, 
@@ -95,30 +86,56 @@ module Data.Metrology.Poly (
   sMOne, sMTwo, sMThree, sMFour, sMFive,
   sSucc, sPred, sNegate,
 
-  -- ** Deprecated synonyms for the ones above
-  pZero, pOne, pTwo, pThree, pFour, pFive,
-  pMOne, pMTwo, pMThree, pMFour, pMFive,
-  pSucc, pPred,
-
   -- * Internal definitions
   -- | The following module is re-exported solely to prevent noise in error messages;
   -- we do not recommend trying to use these definitions in user code.
   module Data.Metrology.Internal
-
   ) where
 
-import Data.Metrology.Z
-import Data.Metrology.Qu
-import Data.Metrology.Dimensions
-import Data.Metrology.Factor
-import Data.Metrology.Units
-import Data.Metrology.Combinators
-import Data.Metrology.LCSU
-import Data.Metrology.Validity
+import Data.Metrology.Poly
+  hiding ( zero, (|+|), (|-|), qSum, qNegate, (*|), (|*), (|/)
+         , numIn, (#), quOf, (%), showIn, convert, constant )
+import Data.Metrology.Qu  ( Qu(Qu) )
 import Data.Metrology.Internal
+import Data.Foldable as F
 import Data.Proxy
+import Data.Metrology.Units
 
-import Data.VectorSpace
+-- | The number 0, polymorphic in its dimension. Use of this will
+-- often require a type annotation.
+zero :: Num n => Qu dimspec l n
+zero = Qu 0
+
+infixl 6 |+|
+-- | Add two compatible quantities
+(|+|) :: (d1 @~ d2, Num n) => Qu d1 l n -> Qu d2 l n -> Qu d1 l n
+(Qu a) |+| (Qu b) = Qu (a + b)
+
+infixl 6 |-|
+-- | Subtract two compatible quantities
+(|-|) :: (d1 @~ d2, Num n) => Qu d1 l n -> Qu d2 l n -> Qu d1 l n
+a |-| b = a |+| qNegate b
+
+-- | Take the sum of a list of quantities
+qSum :: (Foldable f, Num n) => f (Qu d l n) -> Qu d l n
+qSum = F.foldr (|+|) zero
+
+-- | Negate a quantity
+qNegate :: Num n => Qu d l n -> Qu d l n
+qNegate (Qu x) = Qu (negate x)
+
+infixl 7 *| , |* , |/
+-- | Multiply a quantity by a scalar from the left
+(*|) :: Num n => n -> Qu b l n -> Qu (Normalize b) l n
+a *| (Qu b) = Qu (a * b)
+
+-- | Multiply a quantity by a scalar from the right
+(|*) :: Num n => Qu a l n -> n -> Qu (Normalize a) l n
+(Qu a) |* b = Qu (a * b)
+
+-- | Divide a quantity by a scalar
+(|/) :: Fractional n => Qu a l n -> n -> Qu (Normalize a) l n
+(Qu a) |/ b = Qu (a / b)
 
 -- | Extracts a numerical value from a dimensioned quantity, expressed in
 --   the given unit. For example:
@@ -131,19 +148,17 @@ import Data.VectorSpace
 --   > inMeters x = x # Meter   
 numIn :: forall unit dim lcsu n.
          ( ValidDLU dim lcsu unit
-         , VectorSpace n
-         , Fractional (Scalar n) )
+         , Fractional n ) 
       => Qu dim lcsu n -> unit -> n
 numIn (Qu val) u
-  = val ^* fromRational
+  = val * fromRational
              (canonicalConvRatioSpec (Proxy :: Proxy (LookupList dim lcsu))
               / canonicalConvRatio u)
 
 infix 5 #
 -- | Infix synonym for 'numIn'
 (#) :: ( ValidDLU dim lcsu unit
-       , VectorSpace n
-       , Fractional (Scalar n) )
+       , Fractional n )
     => Qu dim lcsu n -> unit -> n
 (#) = numIn
 
@@ -157,47 +172,38 @@ infix 5 #
 --   > height = 2.0 % Meter
 quOf :: forall unit dim lcsu n.
          ( ValidDLU dim lcsu unit
-         , VectorSpace n
-         , Fractional (Scalar n) )
+         , Fractional n )
       => n -> unit -> Qu dim lcsu n
 quOf d u
-  = Qu (d ^* fromRational
+  = Qu (d * fromRational
                (canonicalConvRatio u
                 / canonicalConvRatioSpec (Proxy :: Proxy (LookupList dim lcsu))))
 
 infixr 9 %
 -- | Infix synonym for 'quOf'
 (%) :: ( ValidDLU dim lcsu unit
-       , VectorSpace n
-       , Fractional (Scalar n) )
+       , Fractional n )
     => n -> unit -> Qu dim lcsu n
 (%) = quOf
 
--- | Use this to choose a default LCSU for a dimensioned quantity.
--- The default LCSU uses the 'DefaultUnitOfDim' representation for each
--- dimension.
-defaultLCSU :: Qu dim DefaultLCSU n -> Qu dim DefaultLCSU n
-defaultLCSU = id
-
--- | The number 1, expressed as a unitless dimensioned quantity.
-unity :: Num n => Qu '[] l n
-unity = Qu 1
-
--- | Cast between equivalent dimension within the same CSU.
---  for example [kg m s] and [s m kg]. See the README for more info.
-redim :: (d @~ e) => Qu d l n -> Qu e l n
-redim (Qu x) = Qu x
+infix 1 `showIn`
+-- | Show a dimensioned quantity in a given unit. (The default @Show@
+-- instance always uses units as specified in the LCSU.)
+showIn :: ( ValidDLU dim lcsu unit
+          , Fractional n
+          , Show unit
+          , Show n )
+       => Qu dim lcsu n -> unit -> String
+showIn x u = show (x # u) ++ " " ++ show u
 
 -- | Dimension-keeping cast between different CSUs.
 convert :: forall d l1 l2 n. 
   ( ConvertibleLCSUs d l1 l2
-  , VectorSpace n
-  , Fractional (Scalar n) ) 
+  , Fractional n )
   => Qu d l1 n -> Qu d l2 n
-convert (Qu x) = Qu $ x ^* fromRational (
+convert (Qu x) = Qu $ x * fromRational (
   canonicalConvRatioSpec (Proxy :: Proxy (LookupList d l1))
   / canonicalConvRatioSpec (Proxy :: Proxy (LookupList d l2)))
-
 
 -- | Compute the argument in the @DefaultLCSU@, and present the result as
 -- lcsu-polymorphic dimension-polymorphic value. Named 'constant' because one
@@ -205,23 +211,7 @@ convert (Qu x) = Qu $ x ^* fromRational (
 -- dimension-polymorphic expressions.
 constant :: ( d @~ e
             , ConvertibleLCSUs e DefaultLCSU l
-            , VectorSpace n
-            , Fractional (Scalar n) )
+            , Fractional n )
          => Qu d DefaultLCSU n -> Qu e l n
 constant = convert . redim
 
-infix 1 `showIn`
--- | Show a dimensioned quantity in a given unit. (The default @Show@
--- instance always uses units as specified in the LCSU.)
-showIn :: ( ValidDLU dim lcsu unit
-          , VectorSpace n
-          , Fractional (Scalar n)
-          , Show unit
-          , Show n )
-       => Qu dim lcsu n -> unit -> String
-showIn x u = show (x # u) ++ " " ++ show u
-
--- | The type of unitless dimensioned quantities.
--- This is an instance of @Num@, though Haddock doesn't show it.
--- This is parameterized by an LCSU and a number representation.
-type Count = MkQu_ULN Number